chore: refactor evm chain enum (#290)
* chore: refactor evm chain enum

* chore: remove random files from tsconfig

* chore: fixing exports of EvmChain

* chore: 0.13.7 docs
