+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 - Fix BigNumberUnits utility to adjust for another `fractional component exceeds decimals` edge case
 - Add AddGasToCosmosChain method to AxelarGMPRecoveryAPI
 - Remove singleton pattern that was used in the `getConfigs` function that caused an issue in toggling environments in the AxelarQueryAPI
+- Refactor "EvmChain" enum
 
 ## [0.13.6] - 2023-SEPTEMBER-1
