+++ src/constants/EvmChain.ts
@@ -1,0 +1,19 @@
+export enum EvmChain {
+  ETHEREUM = "ethereum",
+  AVALANCHE = "avalanche",
+  FANTOM = "fantom",
+  POLYGON = "polygon",
+  POLYGON_ZKEVM = "polygon-zkevm",
+  MOONBEAM = "moonbeam",
+  AURORA = "aurora",
+  BINANCE = "binance",
+  BNBCHAIN = "binance",
+  ARBITRUM = "arbitrum",
+  CELO = "celo",
+  KAVA = "kava",
+  BASE = "base",
+  FILECOIN = "filecoin",
+  OPTIMISM = "optimism",
+  LINEA = "linea",
+  MANTLE = "mantle",
+}
