+++ src/libs/TransactionRecoveryApi/AxelarRecoveryApi.ts
@@ -5,7 +5,8 @@ import { loadChains } from "../../chains";
 import { EnvironmentConfigs, getConfigs } from "../../constants";
 import { retry, throwIfInvalidChainIds } from "../../utils";
 import { AxelarQueryClient, AxelarQueryClientType } from "../AxelarQueryClient";
-import { AxelarRecoveryAPIConfig, Environment, EvmChain, EvmWalletDetails } from "../types";
+import { AxelarRecoveryAPIConfig, Environment, EvmWalletDetails } from "../types";
+import { EvmChain } from "../../constants/EvmChain";
 import EVMClient from "./client/EVMClient";
 import { broadcastCosmosTxBytes } from "./client/helpers/cosmos";
 import rpcInfo from "./constants/chain";
