+++ src/libs/test/TransactionRecoveryAPI/AxelarGMPRecoveryAPI.spec.ts
@@ -4,10 +4,10 @@ import {
   ApproveGatewayError,
   AxelarTxResponse,
   Environment,
-  EvmChain,
   EvmWalletDetails,
-  GasToken,
 } from "../../types";
+import { EvmChain } from "../../../constants/EvmChain";
+import { GasToken } from "../../../constants/GasToken";
 import { createNetwork, utils } from "@axelar-network/axelar-local-dev";
 import { Contract, ContractReceipt, ContractTransaction, ethers, Wallet } from "ethers";
 import DistributionExecutable from "../abi/DistributionExecutable.json";
